+++ pkg/analyzer/lib/src/dart/analysis/search.dart
@@ -590,7 +590,7 @@ class Search {
       if (unitResult is ResolvedUnitResult) {
         CompilationUnit unit = unitResult.unit;
         for (Directive directive in unit.directives) {
-          if (directive is PartOfDirective && directive.element == element) {
+          if (directive is PartOfDirective && directive.element2 == element) {
             results.add(
               SearchResult._(
                 unit.declaredElement!,